extracting types
